pyproject.toml
[tool.poetry]
name = "qr-code-generator-api-4"
version = "0.1.0"
description = "Given the information gathered from the user, the application's primary functionality centers around generating QR codes based on user-provided data, which encompasses URLs, textual content, or contact information. To cater to the user's preferences and requirements, the application will allow for customization of the QR code's physical attributes, including its size, with a minimum specification of 2 x 2 cm to ensure clarity and ease of scanning, and color, emphasizing a high contrast between the QR code and its background for optimal readability across various devices and scanning conditions, traditionally adopting a black-on-white configuration. Moreover, the user specifies a medium error correction level to balance data density and integrity effectively, making it adaptable across diverse applications while maintaining a satisfactory level of error correction. Lastly, the user has a preference for the QR code's output format to be in SVG, catering to scalability and high-resolution requirements. The tech stack chosen for implementing this functionality includes Python as the programming language, FastAPI for the API framework to handle requests efficiently, PostgreSQL for the database to store user preferences and generated codes, and Prisma as the ORM for seamless database integration and operations."
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
bcrypt = "^3.2.0"
fastapi = "^0.78.0"
passlib = {version = "*", extras = ["bcrypt"]}
prisma = "*"
pydantic = "*"
python-jose = {version = "*", extras = ["cryptography"]}
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
